pyproject.toml
[tool.poetry]
name = "deep-agent-agi"
version = "0.1.0"
description = "Deep Agent AGI - General purpose deep agent framework with cost optimization"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "deep_agent", from = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
# Core FastAPI
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.0"

# LangChain & DeepAgents (Stable 1.0 versions - Oct 2025)
langchain = "^1.0.0"           # Stable 1.0.2 (Oct 21, 2025)
langchain-core = "^1.0.0"      # Required by DeepAgents 0.2.0
langchain-openai = "^0.3.0"    # Compatible with langchain-core 1.0
langgraph = "^1.0.0"           # Stable 1.0.1 (Oct 20, 2025)
langgraph-checkpoint = "^3.0.0"  # Compatible with LangGraph 1.0+
langgraph-checkpoint-sqlite = "^3.0.0"  # SQLite checkpoint storage
langgraph-prebuilt = "^1.0.0"  # Required by LangGraph 1.0.1
langsmith = "^0.4.0"           # Compatible with LangChain 1.0
deepagents = "^0.2.0"          # Stable 0.2.0 (Oct 28, 2025) - Released TODAY!

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
alembic = "^1.13.0"
asyncpg = "^0.29.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.0"}
pgvector = "^0.3.0"

# MCP & External Integrations
mcp = "^1.0.0"
perplexity-mcp = "^0.1.0"

# Monitoring & Observability
opik = "^1.7.17"  # Updated to be compatible with opik-optimizer (requires >=1.7.17)
opik-optimizer = {version = "^2.2.0", python = ">=3.11,<3.14"}

# Prompt Optimization & Statistical Analysis
scipy = "^1.11.0"
numpy = "^1.24.0"

# Security
bcrypt = "^4.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

# Utilities
httpx = "^0.27.0"
tenacity = "^9.0.0"
structlog = "^24.4.0"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-html = "^4.1.0"
pytest-json-report = "^1.5.0"
pytest-mock = "^3.14.0"

# Linting & Formatting
ruff = "^0.7.0"
mypy = "^1.13.0"

# Pre-commit & Security
pre-commit = "^4.0.0"
bandit = "^1.7.0"
detect-secrets = "^1.5.0"

# Type Stubs
types-requests = "^2.32.0"
types-passlib = "^1.7.0"

# Development Tools
ipython = "^8.29.0"
ipdb = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
cache-dir = ".cache/ruff"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.11"
cache_dir = ".cache/mypy"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langgraph.*",
    "langsmith.*",
    "deepagents.*",
    "opik.*",
    "mcp.*",
    "pgvector.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
cache_dir = ".cache/pytest"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["backend/deep_agent"]
data_file = "out/coverage/.coverage"
omit = [
    "*/tests/*",
    "*/conftest.py",
    "*/__init__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
